gettrends/requirements.txt
pandas==1.2.0
newspaper3k==0.2.8
pyspellchecker==0.5.6
GoogleNews==1.5.2
random-word==1.0.4
nltk==3.5